feat: implement environment variable checks in layout and remove redundant checks from auth
